requirements.txt
# Web Framework
Flask==2.3.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
Jinja2==3.1.2

# Machine Learning & Data Science
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
imbalanced-learn==0.11.0

# Data Processing & Visualization
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2

# Web Deployment & Config
gunicorn==21.2.0
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.0
pytest-cov==4.1.0
